% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insol_functions.R
\name{declination}
\alias{declination}
\title{Solar declination}
\usage{
declination(jd)
}
\arguments{
\item{jd}{Julian Day}
}
\value{
Solar declination in degrees
}
\description{
Solar declination
}
\keyword{internal}
