% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insol_functions.R
\name{degrees}
\alias{degrees}
\title{Convert between radians and degrees}
\usage{
degrees(radian)
}
\arguments{
\item{radian}{Angle in radians}
}
\value{
Angle in degrees
}
\description{
Convert between radians and degrees
}
\keyword{internal}
