% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insol_functions.R
\name{hourangle}
\alias{hourangle}
\title{Hour angle}
\usage{
hourangle(jd, longitude, timezone)
}
\arguments{
\item{jd}{Julian Day}

\item{longitude}{Longitude in degrees}

\item{timezone}{Timezone in hours (west is negative)}
}
\value{
Hour angle in radians
}
\description{
Hour angle
}
\keyword{internal}
