% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insol_functions.R
\name{JD}
\alias{JD}
\title{Julian Day conversion}
\usage{
JD(x, inverse = FALSE)
}
\arguments{
\item{x}{POSIXct time object or Julian Day numeric}

\item{inverse}{Logical, if TRUE convert from JD to POSIXct}
}
\value{
Julian Day (numeric) or POSIXct time
}
\description{
Julian Day conversion
}
\keyword{internal}
