% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insol_functions.R
\name{radians}
\alias{radians}
\title{Convert between degrees and radians}
\usage{
radians(degree)
}
\arguments{
\item{degree}{Angle in degrees}
}
\value{
Angle in radians
}
\description{
Convert between degrees and radians
}
\keyword{internal}
