% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insol_functions.R
\name{sunpos}
\alias{sunpos}
\title{Solar position (azimuth and zenith)}
\usage{
sunpos(sunv)
}
\arguments{
\item{sunv}{Sun vector (3-column matrix from sunvector())}
}
\value{
2-column matrix with azimuth and zenith in degrees
}
\description{
Solar position (azimuth and zenith)
}
\keyword{internal}
