% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insol_functions.R
\name{sunvector}
\alias{sunvector}
\title{Solar vector}
\usage{
sunvector(jd, latitude, longitude, timezone)
}
\arguments{
\item{jd}{Julian Day}

\item{latitude}{Latitude in degrees}

\item{longitude}{Longitude in degrees}

\item{timezone}{Timezone in hours (west is negative)}
}
\value{
3-column matrix with x, y, z coordinates of sun vector
}
\description{
Solar vector
}
\keyword{internal}
